requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
autoflake==2.3.1
    # via -r requirements.in
black==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via
    #   black
    #   pip-tools
customtkinter==5.2.2
    # via -r requirements.in
darkdetect==0.8.0
    # via customtkinter
flake8==7.1.1
    # via -r requirements.in
isort==5.13.2
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
mysql-connector-python==9.3.0
    # via -r requirements.in
numpy==2.2.1
    # via pandas
packaging==24.2
    # via
    #   black
    #   build
    #   customtkinter
pandas==2.2.3
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   black
    #   yamllint
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via
    #   autoflake
    #   flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   yamllint
ruamel-yaml==0.17.40
    # via yamlfmt
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
six==1.17.0
    # via python-dateutil
tzdata==2024.2
    # via pandas
wheel==0.45.1
    # via pip-tools
yamlfmt==1.1.1
    # via -r requirements.in
yamllint==1.35.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
